backend/requirements.txt
# CoreMatch Backend Dependencies
# Python 3.11+

# Web Framework
Flask==3.0.3
flask-cors==4.0.1

# Database
psycopg2-binary==2.9.9

# Auth & Security
PyJWT==2.8.0
bcrypt==4.1.3
cryptography==42.0.8

# Rate Limiting
Flask-Limiter==3.7.0

# Background Jobs
redis==5.0.8
rq==1.16.2

# Cloud Storage (Cloudflare R2 via S3-compatible API)
boto3==1.34.131
botocore==1.34.131

# AI - Groq
groq==0.9.0

# AI - Pipeline providers (optional, for multi-provider support)
anthropic>=0.25.0
openai>=1.30.0

# Video Processing
# ffmpeg-python is a Python wrapper; ffmpeg binary must be installed separately
ffmpeg-python==0.2.0

# Email (AWS SES via SMTP)
# Uses Python's built-in smtplib + email modules

# SMS (Twilio)
twilio==9.2.3

# Monitoring
sentry-sdk[flask]==2.9.0

# Payments
stripe>=8.0.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
requests==2.32.3

# Testing
pytest==8.2.2
pytest-mock==3.14.0
httpx==0.27.0

# Security - Input validation
email-validator==2.2.0

# PDF Generation
fpdf2>=2.7.0

# Document Parsing (CV extraction for AI pipeline)
PyPDF2>=3.0.0
python-docx>=0.8.11

# WSGI server for production
gunicorn==22.0.0
